+++ gamemodes/terrortown/entities/weapons/weapon_ttt_cse.lua
@@ -113,6 +113,12 @@ if CLIENT then
     function SWEP:Initialize()
         self:AddTTT2HUDHelp("vis_help_pri")
 
+        self.BaseClass.Initialize(self)
+    end
+
+    ---
+    -- @realm client
+    function SWEP:InitializeCustomModels()
         self:AddCustomViewModel("vmodel", {
             type = "Model",
             model = "models/Items/battery.mdl",
@@ -142,8 +148,6 @@ if CLIENT then
             skin = 0,
             bodygroup = {},
         })
-
-        self.BaseClass.Initialize(self)
     end
 
     ---